Update unit tests to new controller structure
